backend/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pydantic[email]==2.10.3

# Paystack Integration
requests==2.32.3

# Environment & Config
python-dotenv==1.0.1

# CORS
fastapi-cors==0.0.6

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
